requirements.txt
nba_api
pandas
numpy
scikit-learn
joblib
scipy
shap
torch
plotly
dash
dash-bootstrap-components
requests
beautifulsoup4
lxml
curl_cffi
playwright
sqlalchemy
python-dotenv
pytrends

